chore(examples): getting rid of manual dark mode - removing ts-ignores - pending patch
